target-app/requirements.txt
Flask==2.0.1
Authlib==1.0.1
requests==2.26.0
PyJWT==2.3.0